Xtend-based generator and builder participant for Scope DSL
Adds a builder participant for the Scope DSL which will generate the
scope provider and scope name provider classes as part of the Eclipse
build.

The implementation is using Xtend, but the class names are kept as
similar as possible to the original Xtend1/Xpand implementation. To
avoid naming conflicts the new classes typically have an 'X' suffix. The
old classes will be removed and the new classes renamed in a follow-up
change.

To find the .ext files containing Java extensions a custom ClassLoader
is now created (see ScopingGeneratorUtil), which will be used by
Xtend1's ResourceLoader. This ClassLoader uses the context project's
classpath as computed by JDT. For the context project itself the source
folders will be used, so that the Java builder doesn't have to come
first in the project description.

The MWE2 workflow fragment will now only generate the required Guice
bindings. Its support for generating test stubs has also been removed.

Note that there is still a lot of code cleanups and Javadoc required.
This will come in a follow-up change when the Export builder participant
has been implemented. Also note that the code in the current state
should in any case be easier to maintain than the original Xtend1/Xpand
implementation.

Change-Id: I03517e9da5370119e562539e99e3558d0c960abb