Store *parsed* source text in binary resource
Instead of obtaining the source text from the file system it is better
to get the actual parsed text. This is faster (as the parsed text is
stored in a String) and avoids problems when the source text from the
file system isn't passed 1-to-1 to the parser (this is e.g. the case for
WFD and will probably also apply to wrapped PL/SQL sources).

Change-Id: I81bf22b59ccc61d918277a0d613d496d978bbca2