Extend export DSL to support export extension
Add reference to the grammar for which extension is created and a name.
Generate the necessary extension classes, so extensions need nothing in
addition to .export file. Extension point for the DSL to allow export
contributions for now stays in the DSL implementation. Will be done in a
next step.

For export files that are not defining an extension nothing changed.

Also aligned naming. Now the whole extension to the language is called
LanguageExtensions as it contributes multiple things to the DSL
implementation. One single extension is called a FeatureExtension as it
affects one language feature/service. Intentionally didn't use Service
as a name here. classes with Service as a suffix are meant to be
consumed by extension points within implementations of export or
inference of the host DSL.

Changed lazy initialization to injection.

Change-Id: Ie7afc42a8fe18acd4ba67faa8ca92440b1a5cd30