Add AbstractValidationTest#resourceDiagnostic()
To be able to test for resource diagnostics this commit adds the methods
resourceDiagnostic() and linkingError() to AbstractValidationTest, which
can be used from an Xtend template just like e.g. diagnostic() and
error().

Along these lines we may in the future also want to add a convenience
syntaxError() method.

Also the code in AbstractValidationTest was cleaned up a little bit.