StackOverflowError protection for DelegatingScope
DelegatingScope implemented a very simple StackOverflowError protection
which only worked if a "scopeof" expression was the last in the chain
and delegated to itself.

This commit implements a more generic StackOverflowError protection
which also works in other cases. The logic is very complicated for what
it does, but that is because if implemented in the most straight-forward
manner the error logging itself would again result in a
StackOverflowError being raised!

It should be noted that the error is logged to assist the developer in
addressing the underlying problem with the scope definitions.