#72: Fix problems with InferredModelAssociator
Allow a single source model element to be the primary source for
multiple inferred model elements and also improve the lookup performance
for the primary source of an inferred model element.

These changes to InferredModelAssociator also required corresponding
changes in DirectLinkingResourceStorageLoadable and
DirectLinkingResourceStorageWritable.

Additionally this commit fixes a bug in
DirectLinkingResourceStorageWritable which could cause incorrect
fragments to be computed for the inferred model element associations map
for inherited containment features as EStructuralFeature#getFeatureID()
was called rather than EClass#getFeatureID(EStructuralFeature). This in
turn could lead to ClassCastExceptions in
DirectLinkingResourceStorageLoadable.