#74: Fix bug in DirectLinkingResourceStorageWritable
Fixes URI fragment computation in getURIFragmentPath() for cases when an
object is contained by another object where the containment feature is
inherited from a supertype which isn't the first ("left-most")
supertype.

The same kind of problem could occur in
AbstractSelectorFragmentProvider#computeSelectorFragmentSegment() as it
also uses EStructuralFeature#getFeatureID(), so that is also fixed.