#72: Also use ordered collection for "sourceToInferredModelMap"
Since there are clients which rely on the order of elements in
InferedModelAssociator's "sourceToInferredModelMap" and hence the order
of elements in the Set returned by
IInferredModelAssociations#getInferredModelElements(), this commit
changes InferedModelAssociator to use LinkedLists rather
than a HashSet and LinkedListBasedSet as the values for both
"sourceToInferredModelMap" and "inferredModelToSourceMap".

Using a List rather than a Set improves the performance for
associatePrimary() at the cost of slower getInferredModelElements() and
getSourceModelElements() operations.

DirectLinkingResourceStorageWritable has also been changed to serialize
an ImmutableSet rather than a HashSet to make sure that this insertion
order is maintained.

It might make sense to add a method getPrimaryInferredModelElement() to
IInferredModelAssociations, which would also be in line with Xbase's
IJvmModelAssociations#getPrimaryJvmElement(). This would make it more
explicit that this insertion order is maintained.