Improve performance of model associations in binary resources
Again changes InferredModelAssociator's internal representation of the
model associations. Now the implementation uses ArrayDeques instead of
LinkedLists, as there are more memory-efficient and typically also
perform better (InferredModelAssociator doesn't remove elements in the
middle).

Additionally the serialization of the model associations has been tuned
to use less space (around a factor 7 in uncompressed form) and perform
better (up to 3 times faster serialization and up to 8 times faster
deserialization). Note however that the model associations is only one
of four parts of a binary resource, so the overall improvements are of
course lower.