Always delete binary storage before writing
To be on the safe side the binary storage should always be deleted
first. Otherwise, if an I/O exception occurs during an INCR build the
old storage could still be present.

Also, if the inferred model associations map contains "dangling" objects
(not part of any resource), a warning will be logged instead of an
exception being thrown.