Add associations store to Xtext platform
So far platform was a set of resource descriptions plus issues. Issues
are not completely independent from resource descriptions as they have
to be associated with some resource description in the builder state.

The new concept of associations to derived objects (created by builder
participants) have a similar flavor - they are attached to existing
resource descriptions (only resources that are processed by the builder
may actually produce some derived objects), but are not part of the
resource description (the core part of Xtext builder doesn't care about
them much).

Change-Id: I57455e12986ca16e4119a3e3737e4e82a52d3916