#117: Fix memory leak in AbstractTypeProvider
Making ComputationData static fixes the memory leak as the values then
no longer reference the ThreadLocal object (preventing the garbage
collection).

This commit also improves AbstractPolymorphicScopeProvider by not doing
unnecessary work in getVisibleContainers() which results in quite a lot
of String (and char[]) objects which need to be garbage collected.