Add generator preference to Check for DSL-internal validations
There is a difference in how DSL-internal checks and external SCA checks
are registered and written. Also internal checks typically have no
useful documentation (i.e. compiler errors are self-explaining) when SCA
checks (coding conventions) do usually need a help page (explaining why
the rule is useful).

In this change we add a project preference so we can tell generator
whether it's now doing an SCA plugin or is generating a check for a DSL.
For internal checks docu generation is then disabled.