Derelease retry in indexing if reference resolution fails
DSLs should not rely on indexing order and if needed use other means to
resolve within indexing phase (i.e. include-like concepts over file
system directly).

So we are replacing an old mechanism that retries indexing of a source
if it attempts to resolve a reference in an indexing phase and fails
with a warning.

Now resource descriptions will be written AS IS in case of a failed
linking attempt. A warning log message will be reported, meaning that
DSL-implementers need to do something.