Cleanup: Throw IOException in DirectLinkingResourceStorageLoadable
With the Xtext upgrade it is now possible to throw IOExceptions from
DirectLinkingResourceStorageLoadable. As a result the exception handling
can be simplified.

Additionally also removed the "Skipping loading node model for synthetic
resource" log message which is no longer required, as the source text is
stored as part of the binary storage.

Finally, MonitoredClusteringBuilderState will now register any
invalidated resources as "sources" which mustn't be loaded from binary
storage. This is to make sure the builder doesn't load these sources
from binary storage when building them. Also, in case the resource to
process was already loaded from binary storage, it will be replaced with
a resource loaded from source.