Support for demand-loading node model from binary storage
Adds the capability of demand-loading the node model (and source) to
DirectLinkingResourceStorageLoadable. By default resources are loaded
without the node model when loaded from binary storage. In its place the
implementation will install ProxyCompositeNode adapters, which will
replace itself with the real node model (loaded from the binary storage)
on demand when details from the node model are accessed. For optimal
load performance the node model is stored last in the zip file, so that
it can be completely ignored when applicable.

How a resource is loaded can be controlled using the ResourceLoadMode,
which needs to be specified as a ResourceSet load option at the time the
resource is first loaded. This also allows loading a resource without a
proxied node model.

Additionally a DSL can also bind the default load mode to use in its
injector. This takes precendece over the default but is overridden by
the load mode specified in the ResourceSet's load options (if present).

Lasty this commit adds a utility EObjectUtil#isLazyLinkingProxy() which
can be used to determine if a given object represents an Xtext
lazy-linking proxy. This is intended to be used in situations when
resolving such proxies must be avoided.