#165: Improve performance of InferenceContainer
In the context of resources loaded from binary storage the overhead of
InferenceContainer is minimized by storing the object URI fragments. As
a consequence these don't need to be recomputed after loading the
resource from binary storage. This was particularly expensive because
this relied on an EMF EContentAdapter which was added to every EObject
of the resource. Thus this commit also slightly lowers the memory
footprint.