Fragment for grammar generation with support for semantic predicates (#174)
The following extension is supported:

A datatype grammar rule containing only one ID terminal rule can be annotated with @KeywordRule annotation provided a list of words so only these words can be accepted by this rule.

Example:

   /**
     * @KeywordRule(visible, invisible)
     * /
   VisibleKind returns VisibleKind:
     ID
   ;

The above rule will accept only 'visible' and 'invisible' identifiers. This rule we called a keyword rule because it is intended to replace usages of keywords which shall not be reserved words in the language. Reserved words are words that are not allowed to be used in identifiers.

The above example can therefore replace the following enumeration:

   enum VisibleKind :
       VISIBLE   = "visible"
     | INVISIBLE = "invisible"
   ;

The generator will insert validating semantic predicates in the rule and gated semantic predicates whenever needed in the grammar. Syntax error messages will be adjusted accordingly.

Also the fragment support more generic

   /**
     * @SemanticPredicate(predicateName)
     * /
   Some Rule returns SomeEObject:
     anything
   ;
to insert validating and gated predicates.