Minor performance improvements
Improves the performance of
InferenceContainerImplCustom#addFragmentMapping() by using
InternalEList#addUnique() rather than EList#add(). This is fine because
we already know that the fragment is unique.

PatternAwareEObjectDescriptionLookUp#getExportedObjects() is also
slightly improved by removing the Iterables#isEmpty() check at the
beginning. There is no reason to optimize this case, as it should
typically not occur anyway.