Fix ClassCastException when lazy loading node model from binary
Until now uninstallProxyNodeModel assumed that when lazy loading the
node model from a binary resource, the resource will not have a node
model attached to any EObject.

However, if the AST has been loaded from the binary resource and then
modified, it might be that an EObjects already has a node model, in
these cases, uninstallProxyNode does not need remove any
ProxyCompositeNode from the adapter list of the EObject.