Check memory when performing resource demand loads
If there is not enough memory to load a needed resource -> abort
processing the current one, clear the resource set and try again later.

The new 'isMemoryLimitReached' method has been introduced in
IBuilderResourceLoadStrategy. It returns 'true' when there is not enough
memory.

The retry logic is essentially a revert of a part of
https://github.com/dsldevkit/dsl-devkit/pull/148