Avoid triggering inference when executing container queries
Calls to resource.getContents.get(0) accidentally trigger inference if
the resource was loaded without inference. Even in cases where the
context.eResource() and originalResource are different, we do not need
to get the an EObject from originalResource to find its visible
containers.

- Deprecate the first parameter in
ContainerQuery#execute(EObject/Resource context, Resource
originalResource) as it is never used
- Add AbstractPolymorphicScopeProvider#getVisibleContainers(XtextResource)
as alternative to
AbstractPolymorphicScopeProvider#getVisibleContainers(EObject, Resource)
as the EObject is only relevant if the resource is null.
- Refactor all methods to avoid useless calls to getContents()