Expose the delegate of ForwardingMap used in ResourceDescriptionStrategy
Expose the delegate of ForwardingMap used in the generated code of the
ResourceDescriptionStrategy. This allows consumers of this map to get a
reference to the ImmutableMap. Until now consumers who needed such an
immutable map where creating a new immutable map which was a copy of the
original one. This overhead can be eliminated by exposing the delegate.