Use the name of descriptions which came from the index
Until now to register the names of the linked objects the LinkingService
always looked for the IEObjectDescription within the eResource of the
linked object, then used the qualified name of this description as the
name to register.

This is very inefficient, as for each linked object, an object
descriptor is looked up again within the resource by comparing the URI's
fragment of the object with all the URI's fragments of all objects
within the given resource (until one match is found). The linear search
is very slow because URI#getFragment is an expensive operation, it will
convert the internal representation of the fragment into a String, then
put the string into an EMF String pool, and from them use Sring#intern
before returning the String.

As it happens, for objects which are linked using the index (the Proxy
objects), the descriptor that scope#getSingleElement() returns has
already the exported name of the object as its name, so we can take its
names without looking again for a descriptors from where to extract the
name.

For objects which are not a Proxy, the returned descriptor is not the
one exported to the index, just a descriptor created on the fly for
scoping, and its name is unrelated to the one exported. In this case we
still must fallback to lookup up the resource again by comparing the
URI's fragment.