Use IQualifiedNameProvider if possible
If possible, do not look for the qualified names in the resource
descriptor, asking the IQualifiedNameProvider for the values is much
faster, as the lookup involve very expensive operations (specially the
second one):
- Create a collection with all the resource descriptions of the
referenced resource (only first lookup).
- Iterate through the list every time and compare the fragments.

Asking the IQualifiedNameProvider has the extra advantage that it can
return a name even the target object has not yet been exported, this
happens in the second phase when:
- a resource is loaded on demand
- liking to an inferred object which was not inferred in the pre-linking
phase is done.
- the resource loaded on demand has not yet been processed in the second
phase

under this circumstances the
IResourceDescriptions.getExportedObjectsByObject will not find the
element we have linked right before.