Optimize DefaultCheckImpl.internalValidate
- The 'methodsForType' field is now a usual map instead of a cache with
weak references. The cached MethodWrapper object do not seem to be
extremely big so it makes sense to compute them once and re-use, as the
internalValidate method gets invoked really often.

- Iteration through the list of the methods is now done using simple for
loop, without creation of an iterator

- Computation of 'methodsForType' moved toward the beginning of the
method, so we don't need to create 'State' if there it isn't needed.