Delete IBuilderResourceLoadStrategy
Since the memory parameters aren't really necessary anymore and
effectively 'mayProcessAnotherResource' method always returns 'true',
the IBuilderResourceLoadStrategy interface and all of its
implementations are deleted in this change.

As a result the outer loop in MonitoredClusteringBuilderState.doUpdate
is deleted too, since all the work is now supposed to be done in the
inner loop. ClusterClosedEvent is also deleted.