Upgrade Xtext to 2.18
Upgrade Xtext to 2.18.

The workaround of rewriting the target platform in the Maven build has
been removed because it could be replaced with vanilla functionality of
Tycho.

Another upgrades, like SWTBot, EMF and Tycho has been performed at the
same time.

In addition, some workspace warnings have been fixed:
- Missing @Override on Xtext files
- Remove use of deprecated constructor in favor of
TypeConformanceComputationArgument.DEFAULT
- Fix path of the genmodel files for the check.core and checkcfg.core
plugin-xml files and remove the metamodel from the classpath to get rid
of warnings on the build.properties file.