Set Target Platform if called from Tycho
Set Target Platform if called from Tycho by using the utility
TargetPlatformUtil provided by Xtext. This allows as to remove the
@Ignore tag from one test.

As a context class we use a random class from the
com.avaloq.tools.ddk.xtext.ui plug-in.

In the past we did that in the DDK by using our own class
TargetDefinitionSetup.java, which was removed in
001676e2d3db8551cb1e17017648643a1513efae, as it was wrongly deemed
unnecessary. Instead of reverting the removal of
TargetDefinitionSetup.java, we prefer to use the code provided by Xtext,
as it does the same and this reduces our code footprint.