Fix loading of target platform in the for the Tycho Execution
Fix loading of target platform in the for the Tycho Execution by using a
class from the com.avaloq.tools.ddk.xtext plugin as a context.

This allows TargetPlatformUtil.setTargetPlatform to skip the current
bundle we are testing, otherwise the folder for the target platform will
include the absolute directory of the maven parent project since the
projects are nested in the parent project the result would be that Java
packages of our project will be available twice and Java won't be able
to find our classes leading in compilation errors during our tests.