Fixes for when to load associations from the binary model.
This includes the following specific fixes:
* When the derived state is discarded, clear the association maps before
removing the derived state because the state needs to exist if the maps
have not been loaded until that point.
* If the proxy associations adapter is not on the resource at the point
when the associations are loaded, then it means that the associations
have already been loaded by a different resource, so there is no need to
load them again.