Do not index derived or transient cross-references
Do as the Xtext base implementation and do not index derived or
transient cross-references. Otherwise calculating the reference
descriptions triggers the calculation of derived/transient features even
if no consumer has asked for the derived features to be calculated.

By doing this, we also keep it in closer to
LazyLinkingResource.resolveLazyCrossReference, which also skips this
types of EReferences.