Allow for visible containers depending on the resource
The containerManager allows for container visibility dependent on the
URI, particularly for sources such as test sources that may want to
ignore/circumvent visibility rules, thus we use the containerManager to
find the visible containers for a resource.

This is also consistent with the algorithm used in
isAffected(Collection<Delta>, IResourceDescription,
IResourceDescriptions) within this class.