Replace ContextActionUiTestUtil with DynamicContextActionUiTestUtil.
SWTBotRootMenu and SWTBotMenu have been updated to deal with dynamic
context menus which fixes the issues ContextActionUiTestUtil was created
to address. They do this by waiting for the requested item(s) to appear,
so this change replaces the old context menu util with a new one that
stops the SWTBot from waiting for a timeout if it can tell that there
are no more items to load into the menu.