Provide a name fallback for objects without name
Provide a name fallback based on the Object#toString method in case the
fragment provider does not provide a name for an object.

Otherwise for EObjects without a name, computeHash would compute a hash
consisting only on the eClass of the object, leading to hashes which
cause collisions when put into a HashMap, which lead to very bad
performance of InferenceContainerImplCustom.addFragmentMapping(EObject).

This bad performance is due to the fact that if many objects have the
same hash, the method solves the collision by creating an unique hash by
increasing and integer by one and appending it to the hash that
collides, then check again, all in a loop until the collision is solved,
the algorithm performance degrades quickly if the number of collisions
is high.