Provide a default qualified name
Provide a default qualified name in case none exist. Otherwise if the
name provider does not provide a name for a given EObject, the
DefaultInferredElementFragmentProvider.computeHash(EClass,
QualifiedName) will create a hash which only depends on the EClass, this
hash will cause collisions in maps and leads to bad performance of
InferenceContainerImplCustom.addFragmentMapping(EObject), which loops
over java.util.Map.putIfAbsent(String, EObject) in case of a collision.