Keep Sequence Number Suffix after call to addFragmentMapping
Keep the sequence number suffix of the last iteration after the call to
addFragmentMapping.

If the counter starts each time at one, the while loop will increment
the counter by one and check the fragmentToObjectMap for existence of
the newly generated key until it reaches the last computed index,
always. This algorithm of O(n^2) complexity is a problem since the
existence check on fragmentToObjectMap is relatively expensive. Since we
know that all the previous indexes have already been given, we can start
at the last free index.

In AbstractExportedNameProvider.getFullyQualifiedName(EObject) a small
code cleanup has been performed.