Remove override on StateBasedContainerWithHandle#filterByURI
Remove override on StateBasedContainerWithHandle#filterByURI, as the
implementation is very close to the one in the basis class delivered by
Xtext.

The only difference is that the DDK implementation initializes the
container state when creating the predicate, and saves a null check on
each iteration of the iterator, and measurement cannot confirm this
makes any difference on real case scenarios.