Fix broken unit tests in Eclipse
The Eclipse project config change in PR 474 (commit
de70e081725e08007bcb1fe6491a971e13c9108e) had the consequence that unit
tests failed in Eclipse.
This change fixes the problem by changing the project config back, but
keeping the ability to at least filter the errors on test sources from
the Problems view in Eclipse.

It also includes a minor fix for an error in Eclipse and the addition of
lsp4j to the dslsdk target.