Do not create a DelegatingResourceDescriptionsData without a delegate
Do not create a DelegatingResourceDescriptionsData without a delegate as
this will crash with an NPE whenever a method of the
DelegatingResourceDescriptionsData is called (all of them used the
delegate). Instead, create a NullResourceDescriptionsData.

This prevents crashing on the Language Server when a context is
activated and the document highlight is run.