Fix crashes in the context of a Language Server
* As in the previous change, do not create a ResourceSetAware without a
delegate as this will crash with an NPE whenever any method is called
(all of them used the delegate). Instead, create a
NullResourceDescriptionsData.

* Avoid processing the resource description currently being added to the
index in ResourceDescriptionsUtil.findReferencesToResources. The class
used by the language server for those descriptors
(Indexer$ResolvedResourceDescription) logs an error if getImportedNames
or getReferenceDescriptions is called. The reason of the Xtext project
for this exception is to detect errors
(https://github.com/eclipse/xtext-core/pull/1807#discussion_r786297617):
"It's noteworthy that these descriptions have been created for resources
that have not been fully processed. If they remain in the index after
the build, subsequent changes cannot correctly detect the affected
resources and the index state will gradually erode and become different
from the actual project contents on disk.
The ResolvedDescriptions should be local to a single build run and never
survive that single build cycle. Within the build loop, there imported
names and reference descriptions should never be necessary."

* Inline with this reasoning, we bring
FixedCopiedResourceDescription closer to the Xtext class
CopiedResourceDescription and log the exceptional case whenever
possible.