Migrate away from legacy Xtext generator
The old CustomClassEcoreGenerator and related classes have been removed.
The CustomClassAwareEcoreGenerator has been improved to respect the Java
Formatter settings regarding indentation by correctly initializing the
JControlModel. This requires overriding the invoke method and proxying a
couple of setters to have access to some of the needed parameters.

Since there is no fragment for the old formatter (only for formatter2,
the languages now bind manually the formatter in its runtime module
until they are migrated to format2 API.

The plugin-gen.xml sources changed the order of it entries and some
whitespace characters, since these changes do not have a functional
effect, the plugin.xml files have not been modified to avoid additional
noise when comparing them.