Add annotation awareness to content assist parser generators.
Adds an AnnotationAwareAntlrContentAssistGrammarGenerator with a common
base class to the AnnotationAwareAntlrGenerator. Updates the generator
fragment to create the content assist parser in the ui plugin if the ide
plugin is not available.

Also updates HelloWorld test grammar so that content assist for keyword
annotated rules and enum rules can be demonstrated.