Add LSP content assist for CheckCfg configuration parameters.
Use the service loader to find configuration parameters as the extension
point registry does not exist in the LSP server.
Filter out dummy java classes from proposals for configuration parameter
values.

Also contains improvement to CheckCfg build performance and a fix to
allow markers to be sent back to the client.