Mr QA: partial revert of PR 771
In https://github.com/dsldevkit/dsl-devkit/pull/771, we removed the
calls to resource.getContents() and resource.eAdapters() in the
exception handler of the loadIntoResource methods.
The XText bug which was the reason for having those 2 calls is closed,
however, it turns out that removing them causes test failures in at
least one consumer. Therefore, this PR re-introduces them pending a
fuller investigation.