Fix the behavior of awaitBinaryStorageExecutorTermination
Previously, `awaitBinaryStorageExecutorQuiescence` did not correctly
attempt to complete pending work. Instead, it was possible for it to
terminate too early (i.e., before the specified timeout), leaving
pending work undone.

To fix this behavior, this commit contains the following changes:

1. Instead of relying on quiescense, the function now explicitly shuts
down the `binaryStorageExecutor` pool to prevent additional work from
being accepted, and attempts to await the termination of actively
running jobs in the pool.
2. Adding work to the pool is now guarded, to account for the
possibility that the pool may be in the "shutdown" state.

As a result of these modifications, the function has been renamed to the
more accurate `awaitBinaryStorageExecutorTermination`.