Do not use a live view of the ExtensionFactoryMap (#857)
* Do not use a live view of the ExtensionFactoryMap

as this means that each time the extensions are accesed, the transform method will be applied. This has a high cost when the extension map is used for all the sources in the workspace. Since the registered extensions do not change once the system is setup, it is much more efficient to map the keys only once.

* Update com.avaloq.tools.ddk.xtext/src/com/avaloq/tools/ddk/xtext/resource/ResourceServiceProviderUtil.java

Co-authored-by: João Dinis Ferreira <hello@joaof.eu>

---------

Co-authored-by: João Dinis Ferreira <hello@joaof.eu>