feat: generate check catalog dispatch methods
That is, modify the Check language generator so that for each check
catalog a dispatch method gets generated for calling all the applicable
check methods for a given check mode and context object.

The motivation for this is to avoid the need to use runtime
introspection to determine what methods are available, and also to avoid
the need to deal with reified Method objects for invocation.

Retain DefaultCheckImpl to avoid breakage, but switch to
DispatchingCheckImpl class as the new default check catalog base class
used by the modified generator.