feat: add AbstractDispatchingCheckImpl base class
That is, create a variant of DispatchingCheckImpl as a slightly
different check catalog implementation base class, one that assumes that
the necessary validation state is passed around across 'validate' and
context method calls as a reference to a DiagnosticCollector. Assume
that instead of using a thread-local variable to track State, as is done
by DispatchingCheckImpl.

To support implementation of the new CheckImpl base class do further
refactoring of the existing ones.

A notable difference of AbstractDispatchingCheckImpl is that it no
longer implements the ValidationMessageAcceptor interface, and it is
instead the DiagnosticCollector that does.