feat: pass DiagnosticCollector to generated methods
That is, instead of using getMessageAcceptor() to access the object for
recording diagnostics, pass a DiagnosticCollector object as an argument
to generated context methods.

Additionally, have the generated calls to 'validate' also specify the
CheckType of the performed check, so that the type can also be included
in any constructed diagnostics. This is something that is missing from
DispatchingCheckImpl based check execution, which always leaves
CheckType as null.