Double-checked locking to reduce overhead of lazy-initialization
The synchronization on the methods setPlatform and basicGetPlatform have
been removed as well, as it is not sensible to have a synchronized
setter without a synchronized getter.

Such synchronization would in any case not prevent usage of old
retrieved platforms. To prevent such cases, where the retrieved platform
is cached locally, consumers need to register as listeners for
notifications.