fix: another thread safety issue in MapCache
In 2025d7584f1ee2a9e5aff796d606806cd9a07273, we improved the thread
safety of the MapCache class by implementing all methods of the Map
interface in the class itself, instead of inheriting the
(non-thread-safe) default implementations of the interface. On closer
inspection, it turns out that the putIfAbsent method was already
implemented, but in a non-thread-safe way.